Swap gnunet-config's default behaviour for the rewrite flag
With the previous default, a configuration file could keep values different
from the defaults even when the user did not explicitly edit that option,
potentially leading to buggy behaviour.

For example: GNUnet's version X+1 changes the default value for a certain
option A, but anyone who has edited the configuration file with version X or
earlier, would still have got the old default for A even when updating to
version X+1.

It was possible to write only the edited parts, but that required explicitly
passing the `--rewrite' (or `-w') flag.

The default behaviour has now been swapped so that the resulting file contains
only differences, while a "frozen" configuration is generated with the
`--rewrite' flag.

Also, as it's a minor change: a function used internally by the logging
component was using translated strings to check the requested log level.  This
behaviour is buggy as passing an untranslated string to
e.g. `GNUNET_log_setup', while the current locale is different and a
translation for that string exists, would generate a different log level than
the one requested.