[FCFSD] Provide a better user experience
The motivations behind these changes are as following.

To begin with, at the most superficial level, the form is given a better
appearance, instead of some plain XHTML.  Additionally, the served pages can
be substituted with something else by using an entry in the configuration
value, altough with some limitations.

The page listing all the registered zones has been removed in favour of a
search function.  A configuration entry could've been used to let service
operators choose between showing the full listing or not, but at the same
time, being presented with a (possibly) giant list of names is not that great
from a usability point of view. Having a search function is, at the very
least, faster than having to wait for the full list to be displayed before
being able to use the user agent's page search feature.

Other than the above, people registering names with the service might not want
to be known by everyone. Even though checking if a certain name or key was
registered already can be known simply by querying the service, it's not
straightforward to associate a name with a specific key (or viceversa).

Last but not least, the service was restructured to be more "route-oriented"
instead of the traditional (X)HTML document format.  The main purpose of this
change is to decouple usage of the service from the tools used to access it.

With a traditional document, users are pretty much forced to use a web
browsers as data submission is carried through the standard HTML form
handling.  Now, it is possible to access the service using any tool capable of
speaking HTTP, regardless of wether it's a web browser, cURL or even a custom
tool specific for this service.

Another advantage of this approach is that it allows adding "layers" to the
service, for example an authentication check before letting users register a
name.  As long as the layer immediately on top of the service is able to send
some JSON using HTTP, there is no need to have users access the service
itself: just put a "proxy" inbetween and run the service locally, while the
proxy handles other administrative tasks before a name can be registered.

By using layers, the service can keep being small feature-wise (i.e. provide
only searching and registering), while everything else is provided by other
applications, including access through protocols other than HTTP.