madmurphy writes:
On AUR I have created the gnunet-uri.desktop file (which is installed as /usr/share/applications/gnunet-uri.desktop) to add support for the gnunet:// URI scheme (so basically it becomes possible to click on gnunet:// URIs in any browser  GNUnet GTK needs to be installed too, but the URI handler gnunet-uri is part of the gnunet package).

You can find a patch attached, which adds

    contrib/xdg-scheme-handler/gnunet-uri.desktop
    contrib/xdg-scheme-handler/README.md

to the git repository.